requirements.txt
colorlog==2.7.0
ipython==5.1.0
beautifulsoup4==4.5.1
progressbar2==3.12.0
Requests==2.12.4
